+++ lib/mongoid/includes.rb
@@ -1,3 +1,4 @@
+require 'mongoid/includes/version'
 require 'mongoid/includes/criteria'
 require 'mongoid/includes/eager_load'
 require 'mongoid/includes/relations/eager'