+++ lib/mongoid/includes/inclusions.rb
@@ -12,15 +12,18 @@ module Mongoid
       end
 
       # Public: Adds a new relation as an inclusion.
+      #
+      # Returns the added inclusion.
       def push(metadata, options = {})
         metadata = Inclusion.new(metadata, options) unless metadata.is_a?(Inclusion)
         super(metadata)
+        metadata
       end
 
       # Public: Checks if the collection already has an inclusion with the
       # specified metadata.
       def include?(metadata)
-        any? { |inclusion| inclusion.metadata == metadata }
+        find { |inclusion| inclusion.metadata == metadata }
       end
     end
   end