+++ spec/support/models/album.rb
@@ -2,6 +2,7 @@ class Album
   include Mongoid::Document
 
   field :name, type: String
+  field :release, type: Date
 
   belongs_to :owner, polymorphic: true
   has_many :songs