+++ spec/support/models/artist.rb
@@ -1,0 +1,8 @@
+class Artist
+  include Mongoid::Document
+
+  field :name, type: String
+
+  has_many :albums
+  has_one :musician
+end