+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## Mongoid::Includes 1.0.0 (July 30, 2015) ##
 
 *   Initial Version.
+
+## Mongoid::Includes 1.1.0 (August 5, 2015) ##
+
+*   Fix error messages for polymorphic includes.
+*   Add :with option that receives the criteria that will be used to include documents.